Correction for GCC compatibility.
git-svn-id: http://b035a0a.online-server.cloud/svn/Acf@627 5144ca80-2444-0410-90ef-8152f95a0219

+++ Include/iqtgui/CLogGuiComp.cpp
@@ -41,7 +41,7 @@ QTreeWidgetItem* CLogGuiComp::CreateGuiItem(const ibase::IMessage& message)
 	QTreeWidgetItem* treeItemPtr = new QTreeWidgetItem;
 	if (treeItemPtr != NULL){
 		QDateTime dateTime;
-		dateTime = QDateTime::fromTime_t(message.GetTimeStamp().ToCTime());
+		dateTime = QDateTime::fromTime_t(uint(message.GetTimeStamp().ToCTime()));
 
 		QString date = dateTime.toString();
 		treeItemPtr->setText(TimeColumn, date);
